Made changes to JMDT explorer URL and ICON (#1934)
* Add JMDT Mainnet chain configuration

Introduces JMDT Mainnet details in additionalChainRegistry and registers chainId 7000700 in chainIds.js for network support.

* Update JMDT Mainnet chain registry export and icons

Changed the file to export the chain data as a named export. Updated infoURL and icon fields to use new URLs, and replaced icon references  for both the chain and explorer.

* Update JMDT chain icon value

Changed the 'icon' field in the JMDT chain registry from a URL to the string 'jmdt' for consistency with other chain icon definitions.

---------

Co-authored-by: Real Shaman <85087525+realdealshaman@users.noreply.github.com>