Update chainid-721529 (name,RPC)
Dear ChainList Team,

We previously submitted a request to change the name of this network but did not receive a response. To resolve the matter, we have now submitted a new request through the company’s official email.

As context, we initially introduced this chain under the name ERAM, but as part of our public launch and in full alignment with our whitepaper and international branding strategy, we have rebranded it to OIL Smart Chain. This change is limited to the name and domain—our technical infrastructure remains unchanged. The purpose of this move is to ensure consistency across the ecosystem and to communicate a clear and reliable message to our users, investors, and global partners.

We kindly request your assistance in reviewing and processing this change at your earliest convenience.

Best regards,Oil Smart Chain Team