Update R0AR Chain Mainnet & Testnet Icons to New Permaweb SVG Location (#2325)
* Add and Update R0AR Chain configuration and Icon

Updating the R0AR Chain configuration found on chainlist. And adding it with Icon for the first time. Chainlist already has the RPC, explorer, name and gas token correct, but this entry was not in the repo and the icon is missing from Chainlist.

* Add configuration and icon via ipfs for R0AR Testnet override

Add r0ar chain icon via ipfs (mainnet + testnet overrides). Chainlist has the correct information already, however, the icon is not designated. This is an override for the information and icon to be corrected for the testnet

* Change default export to named export for chain data

* Update chainid-193939.js

* Update icon URL and format in chainid-193939.js

We are changing our icon link to the permaweb location as well as the SVG from PNG format.

* Update icon URL and format in chainid-11166111.js

We are updating to the new permaweb location for the icon and changing format from png to svg.

---------

Co-authored-by: Define101 <93603962+Define101@users.noreply.github.com>